[Booking][Run] Moving histogram name setting
Instead of defining fake columns in the Run part, the histogram name is set
automatically (even if the Histogram object already has a name, this can
eventually be changed) when the Unit constructor is called.
This is done by creating a new instantiation of the action. Note that
Histogram does not take anymore "name" as parameter from the user side,
since it will be created automatically.